# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  ads-mrkt_internal_market_application_market_http_model.CreateDealRequest:
    properties:
      channel_id:
        type: integer
      details:
        items:
          type: integer
        type: array
      duration:
        type: integer
      listing_id:
        type: integer
      message:
        type: string
      price:
        type: number
      type:
        type: string
    type: object
...
  ads-mrkt_internal_market_application_market_http_model.DealChatLinkResponse:
    properties:
      chat_link:
        type: string
    type: object
  ads-mrkt_internal_market_application_market_http_model.DealResponse:
    properties:
      channel_id:
        type: integer
      created_at:
        type: string
      details:
        items:
          type: integer
        type: array
      duration:
        type: integer
      escrow_address:
        type: string
      escrow_amount:
        type: integer
      escrow_release_time:
        type: string
      id:
        type: integer
      lessee_id:
        type: integer
      lessee_payout_address:
        type: string
      lessee_signature:
        type: string
      lessor_id:
        type: integer
      lessor_payout_address:
        type: string
      lessor_signature:
        type: string
      listing_id:
        type: integer
      message:
        type: string
      price:
        type: number
      status:
        $ref: '#/definitions/ads-mrkt_internal_market_domain_entity.DealStatus'
      type:
        type: string
      updated_at:
        type: string
    type: object
  ads-mrkt_internal_market_application_market_http_model.SetDealPayoutRequest:
    properties:
      wallet_address:
        type: string
    type: object
...
  ads-mrkt_internal_market_application_market_http_model.UpdateDealDraftRequest:
    properties:
      details:
        items:
          type: integer
        type: array
      duration:
        type: integer
      price:
        type: number
      type:
        type: string
    type: object
...
  ads-mrkt_internal_market_domain_entity.DealStatus:
    enum:
    - draft
    - approved
    - waiting_escrow_deposit
    - escrow_deposit_confirmed
    - in_progress
    - waiting_escrow_release
    - escrow_release_confirmed
    - completed
    - waiting_escrow_refund
    - escrow_refund_confirmed
    - expired
    - rejected
    type: string
    x-enum-varnames:
    - DealStatusDraft
    - DealStatusApproved
    - DealStatusWaitingEscrowDeposit
    - DealStatusEscrowDepositConfirmed
    - DealStatusInProgress
    - DealStatusWaitingEscrowRelease
    - DealStatusEscrowReleaseConfirmed
    - DealStatusCompleted
    - DealStatusWaitingEscrowRefund
    - DealStatusEscrowRefundConfirmed
    - DealStatusExpired
    - DealStatusRejected
...
  ads-mrkt_internal_server_templates_response.Template:
    properties:
      data: {}
      error_code:
        type: string
      ok:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Swagger for Ads Market API
  version: "1.0"
paths:
...
  /market/deals:
    post:
      consumes:
      - application/json
      parameters:
      - description: deal body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ads-mrkt_internal_market_application_market_http_model.CreateDealRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Created deal
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  $ref: '#/definitions/ads-mrkt_internal_market_application_market_http_model.DealResponse'
              type: object
        "400":
          description: Bad request
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Not found
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
      security:
      - JWT: []
      summary: Create deal
      tags:
      - Market
  /market/deals/{id}:
    get:
      parameters:
      - description: Deal ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Deal
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  $ref: '#/definitions/ads-mrkt_internal_market_application_market_http_model.DealResponse'
              type: object
        "400":
          description: Bad request
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Not found
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
      security:
      - JWT: []
      summary: Get deal by ID (only if caller is lessor or lessee)
      tags:
      - Market
    patch:
      consumes:
      - application/json
      parameters:
      - description: Deal ID
        in: path
        name: id
        required: true
        type: integer
      - description: fields to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ads-mrkt_internal_market_application_market_http_model.UpdateDealDraftRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated deal
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  $ref: '#/definitions/ads-mrkt_internal_market_application_market_http_model.DealResponse'
              type: object
        "400":
          description: Bad request
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Not found
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
      security:
      - JWT: []
      summary: Update deal draft (type, duration, price, details). Clears both signatures.
      tags:
      - Market
  /market/deals/{id}/chat-link:
    post:
      parameters:
      - description: Deal ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Chat link to open in Telegram
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  $ref: '#/definitions/ads-mrkt_internal_market_application_market_http_model.DealChatLinkResponse'
              type: object
        "400":
          description: Bad request
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Not found
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
      security:
      - JWT: []
      summary: Get deal forum chat link. Caller must be lessor or lessee.
      tags:
      - Market
  /market/deals/{id}/payout-address:
    put:
      consumes:
      - application/json
      parameters:
      - description: Deal ID
        in: path
        name: id
        required: true
        type: integer
      - description: wallet_address (raw)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ads-mrkt_internal_market_application_market_http_model.SetDealPayoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated deal
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  $ref: '#/definitions/ads-mrkt_internal_market_application_market_http_model.DealResponse'
              type: object
        "400":
          description: Bad request
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Not found
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
      security:
      - JWT: []
      summary: Set your payout address on the deal (lessor or lessee). Required before
        signing. Draft only.
      tags:
      - Market
  /market/deals/{id}/reject:
    post:
      parameters:
      - description: Deal ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Updated deal
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  $ref: '#/definitions/ads-mrkt_internal_market_application_market_http_model.DealResponse'
              type: object
        "400":
          description: Bad request
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
        "403":
          description: Forbidden (deal not draft or not a side)
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Not found
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
      security:
      - JWT: []
      summary: Reject deal. Only allowed when deal status is draft; caller must be
        lessor or lessee.
      tags:
      - Market
  /market/deals/{id}/sign:
    post:
      parameters:
      - description: Deal ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Deal (possibly approved)
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  $ref: '#/definitions/ads-mrkt_internal_market_application_market_http_model.DealResponse'
              type: object
        "400":
          description: Bad request
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Not found
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
      security:
      - JWT: []
      summary: Sign deal (lessor or lessee). When both have signed same terms, status
        becomes approved.
      tags:
      - Market
...
  /market/listings/{listing_id}/deals:
    get:
      parameters:
      - description: Listing ID
        in: path
        name: listing_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of deals
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/ads-mrkt_internal_market_application_market_http_model.DealResponse'
                  type: array
              type: object
        "400":
          description: Bad request
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
      security:
      - JWT: []
      summary: List deals by listing ID (only deals where caller is lessor or lessee)
      tags:
      - Market
...
  /market/my-deals:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: List of user's deals
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/ads-mrkt_internal_market_application_market_http_model.DealResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
      security:
      - JWT: []
      summary: List deals for the current user (as lessor or lessee)
      tags:
      - Market
...
schemes:
- http
security:
- JWT: []
securityDefinitions:
  JWT:
    description: Bearer token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"